Crate symbolic_expressions

symbolic-expressions parsing and generating library

Re-exports

pub use iteratom::from_sexp;

Modules

iteratom
high-level API for deconstructing symbolic-expressions

parser
symbolic-expression parser code: data -> symbolic-expression

ser
symbolic-expression serialization code: symbolic-expression -> data

Structs

ParseError
detailed symbolic-expression parse error information

Enums

Sexp
a symbolic-expression structure Can be a string or a list or nothing

SexpError
errors that can happen in this library

Traits

Formatter
trait for formatting the serialization of a symbolic-expression

IntoSexp
like Into trait but works from a ref avoiding consumption or expensive clone

Functions

encode_string
encode a string according to the guidelines given by Kicad

parse_error
utility function that creates a symbolic-expressions Error Result for a parser error

Type Definitions

Rules
datatype used by the example Formatter RulesFormatter