Crate symbolic_expressions

symbolic-expressions parsing and generating library

Modules

parser
symbolic-expression parser code: data -> symbolic-expression

ser
symbolic-expression serialization code: symbolic-expression -> data

Structs

ParseError
detailed symbolic-expression parse error information

Enums

Error
Error type for symbolic-expressions

Sexp
a symbolic-expression Can be a string or a list or nothing

Traits

Formatter
trait for formatting the serialization of a symbolic-expression

IntoSexp
like Into trait but works from a ref avoiding consumption or expensive clone

Functions

encode_string
encode a string according to the guidelines given by Kicad

parse_error
utility function that creates a symbolic-expressions Error Result for a parser error

str_error
utility function that creates a symbolic-expressions Error Result from a String

Type Definitions

Result
symbolic-expressions Result type

Rules
datatype used by the example Formatter RulesFormatter